start to remove config and move all visualizations in the same folder
