+++ src/scipp/__init__.py
@@ -51,14 +51,12 @@ from . import geometry
 # Import functions
 
 # Import python functions
-from .show import show, make_svg
-
-from .html import to_html, make_html, table
+from .visualizations import show, make_svg, to_html, make_html, table
 
 setattr(Variable, '_repr_html_', make_html)
 setattr(DataArray, '_repr_html_', make_html)
 setattr(Dataset, '_repr_html_', make_html)
-del html
+del visualizations
 
 from .io.hdf5 import save_hdf5 as _save_hdf5
