--- src/scipp/configuration/config_default.yaml
@@ -1,28 +1,0 @@
-## Configuration file for scipp
-## See https://scipp.github.io/reference/runtime-configuration.html
-##
-## COLORS
-## All colors are given as one of
-##     - a hex string, such as #ff00ff
-##     - a scalar grayscale intensity such as 0.75 (as a string)
-##     - a legal html color name, e.g., red, blue, darkslategray
-##
-
-# Colors used by output to Jupyter notebooks by, e.g. show, table, plot
-colors:
-  attrs: '#ff5555'
-  coords: '#c6e590'
-  data: '#f6d028'
-  masks: '#c8c8c8'
-
-  # Color for plot control button.
-  button: '#bdbdbd49'
-  # Color for selected plot control button.
-  button_selected: '#bdbdbdbb'
-  # Color for table header text.
-  header_text: '#111111'
-  # Color for hovering on table rows.
-  hover: '#d6eaf8'
-
-# Maximum number of rows in a table to display at the same time.
-table_max_size: 50
