+++ src/scipp/visualizations/__init__.py
@@ -1,0 +1,9 @@
+# SPDX-License-Identifier: BSD-3-Clause
+# Copyright (c) 2023 Scipp contributors (https://github.com/scipp)
+# @file
+
+from .html import to_html, make_html
+from .show import show, make_svg
+from .table import table
+
+__all__ = ["html", "show", "table"]
