+++ src/scipp/visualizations/formatting_datagroup_html.py
