+++ src/scipp/visualizations/formatting_html.py
