+++ src/scipp/visualizations/html.py
