+++ src/scipp/visualizations/resources.py
