+++ src/scipp/visualizations/table.py
