+++ src/scipp/visualizations/templates/__init__.py
