+++ src/scipp/visualizations/templates/datagroup.css
