+++ src/scipp/visualizations/templates/dg_atomic_row.html
