+++ src/scipp/visualizations/templates/dg_collapsible_row.html
