+++ src/scipp/visualizations/templates/dg_detail_list.html
