+++ src/scipp/visualizations/templates/dg_repr.html
