+++ src/scipp/visualizations/templates/icons-svg-inline.html
