+++ src/scipp/visualizations/templates/style.css.template
